+++ examples/data-sources/iosxe_policy_map/data-source.tf
@@ -1,3 +1,3 @@
 data "iosxe_policy_map" "example" {
-  name = "dot1x_policy"
+  name = "POLICY1"
 }