+++ examples/resources/iosxe_policy_map/import.sh
@@ -1,1 +1,1 @@
-terraform import iosxe_policy_map.example "Cisco-IOS-XE-native:native/policy/Cisco-IOS-XE-policy:policy-map=dot1x_policy"
+terraform import iosxe_policy_map.example "Cisco-IOS-XE-native:native/policy/Cisco-IOS-XE-policy:policy-map=POLICY1"